fix: remove dynamic discovery of templates
