+++ install.yml
@@ -1,0 +1,9 @@
+$schema: http://json-schema.org/draft-07/schema
+$id: https://adobe.io/schemas/app-builder-templates/1
+ 
+categories:
+  - code
+  - ui
+ 
+extension:
+  serviceCode: dx/excshell/1
